infra/requirements.txt
cdktf~=0.20.10
cdktf-cdktf-provider-azurerm~=13.11.0
python-dotenv==1.0.1